+++ Backend/src/P2Project.Application/Volunteers/UpdatePhoneNumbers/UpdatePhoneNumbersCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace P2Project.Application.Volunteers.UpdatePhoneNumbers
+{
+    public record UpdatePhoneNumbersCommand(
+        Guid VolunteerId,
+        UpdatePhoneNumbersDto PhoneNumbersDto);
+}
