+++ Backend/src/P2Project.Application/Volunteers/UpdatePhoneNumbers/UpdatePhoneNumbersDto.cs
@@ -1,0 +1,7 @@
+﻿using P2Project.Application.Dtos;
+
+namespace P2Project.Application.Volunteers.UpdatePhoneNumbers
+{
+    public record UpdatePhoneNumbersDto(
+        IEnumerable<PhoneNumberDto> PhoneNumbers);
+}
